Split the Axon Server WebSecurity init into a Configuration and a Configurer, with a pluggable Configurer.
